[WIP] magit-insert-{local,remote}-branches: use for-each-ref
9cac1f53 (magit-insert-{local,remote}-branches: use NUL separated
fields, 2017-10-11) breaks magit-show-refs because "git branch" does
not support the --format option in Git versions earlier than 2.13.

[WIP] We can *almost* use for-each-ref as a drop-in replacement.  The
only discrepancy I've noticed is that the local refs do not include an
entry for a detached head.  At first glance, I don't see an issue with
dropping this because the buffer is about listing refs.  Thoughts?

Fixes #3203.